+++ contrib/devtools/iwyu/bitcoin.core.imp
@@ -3,4 +3,5 @@
   { include: [ "<bits/termios-c_lflag.h>", private, "<termios.h>", public ] },
   { include: [ "<bits/termios-struct.h>", private, "<termios.h>", public ] },
   { include: [ "<bits/termios-tcflow.h>", private, "<termios.h>", public ] },
+  { include: [ "<bits/chrono.h>", private, "<chrono>", public ] },
 ]
